move out in-memory backend to rustic_testing
