+++ crates/core/Cargo.toml
@@ -142,6 +142,7 @@ rstest = { workspace = true }
 rustdoc-json = "0.8.9"
 # We need to have rustic_backend here, because the doc-tests in lib.rs of rustic_core
 rustic_backend = { workspace = true }
+rustic_testing = { workspace = true }
 rustup-toolchain = "0.1.6"
 simplelog = "0.12.2"
 tar = "0.4.40"
