+++ crates/testing/Cargo.toml
@@ -6,7 +6,11 @@ publish = false
 
 [dependencies]
 aho-corasick = { workspace = true }
+anyhow = "1.0.81"
+bytes = "1.5.0"
+enum-map = "2.7.3"
 once_cell = "1.19.0"
+rustic_core = { workspace = true }
 tempfile = { workspace = true }
 
 [lints]
